+++ Changelog.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## [0.1.5] - 2023-02-23
+
 ## Added
 
 - Implemented `SizeOf` for `std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV6}`
@@ -59,7 +61,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for the [`chrono`](https://docs.rs/chrono) crate under the `chrono` feature
 
 <!-- next-url -->
-[Unreleased]: https://github.com/Kixiron/size-of/compare/v0.1.4...HEAD
+[Unreleased]: https://github.com/Kixiron/size-of/compare/v0.1.5...HEAD
+[0.1.5]: https://github.com/Kixiron/size-of/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/Kixiron/size-of/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/Kixiron/size-of/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/Kixiron/size-of/compare/...v0.1.2